requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
lxml==4.7.1
multitasking==0.0.10
numpy==1.22.1
pandas==1.4.0
praw==7.5.0
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
yfinance==0.1.69
